requirements.txt
# ============================================================
# AI-Based Hearing Loss Prediction System - Requirements
# ============================================================

# Core ML / Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0
scipy>=1.7.0

# Audio Feature Extraction
librosa>=0.9.0
soundfile>=0.10.0

# Database
mysql-connector-python>=8.0.0

# API Framework
Flask>=2.0.0
flask-cors>=3.0.0
Werkzeug>=2.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Utilities
python-dotenv>=0.19.0

# Development / Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing
pytest>=7.0.0
pytest-flask>=1.2.0
